requirements.txt
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
DateTime==4.3
forex-python==1.5
greenlet==1.0.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.4.1
numpy==1.20.2
pandas==1.2.4
Pillow==8.2.0
psycopg2==2.8.6
pyodbc==4.0.30
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.4.11
tqdm==4.60.0
urllib3==1.26.4
zope.interface==5.4.0
